The Readditon of Express and the Additon of Mongoose
On my local computer I deleted the node_modules file, because I thought
there were extra files that the package.json didnt not call for. When I
rebuilt the project all of those files were added again. This makes them
necessary.

I also added Mongoose so we have a way to interact with the MongoDB